dataflow-types: Protobuf definition for plan
Some refactoring was needed to avoid import cycles
