+++ src/dataflow-types/build.rs
@@ -12,6 +12,7 @@ fn main() {
         .extern_path(".mz_expr.linear", "::mz_expr")
         .extern_path(".mz_expr.relation", "::mz_expr")
         .extern_path(".mz_expr.scalar", "::mz_expr")
+        .extern_path(".mz_expr.id", "::mz_expr")
         .extern_path(".mz_repr.global_id", "::mz_repr::global_id")
         .extern_path(".mz_repr.proto", "::mz_repr::proto")
         .extern_path(".mz_repr.row", "::mz_repr")
