+++ src/dataflow-types/src/plan/join.proto
@@ -11,7 +11,6 @@
 
 syntax = "proto3";
 
-import "google/protobuf/empty.proto";
 import "expr/src/linear.proto";
 import "expr/src/scalar.proto";
