+++ src/dataflow-types/src/plan/reduce.proto
@@ -11,10 +11,10 @@
 
 syntax = "proto3";
 
-import "google/protobuf/empty.proto";
 import "expr/src/linear.proto";
 import "expr/src/relation.proto";
 import "expr/src/scalar.proto";
+import "google/protobuf/empty.proto";
 
 package mz_dataflow_types.plan.reduce;
